backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
psycopg2-binary==2.9.9
openpyxl==3.1.5
django-cors-headers==4.4.0
